+++ R/std.R
@@ -1,7 +1,11 @@
 #' @title Standardize and center variables
 #' @name std
+#'
 #' @description \code{std()} computes a z-transformation (standardized and centered)
-#'              on the input. \code{center()} centers the input.
+#'   on the input. \code{center()} centers the input. \code{std_if()} and
+#'   \code{center_if()} are scoped variants of \code{std()} and \code{center()},
+#'   where transformation will be applied only to those variables that match the
+#'   logical condition of \code{predicate}.
 #'
 #' @param include.fac Logical, if \code{TRUE}, factors will be converted to numeric
 #'   vectors and also standardized or centered.
@@ -12,22 +16,40 @@
 #'   absolute deviation (MAD, \code{robust = "mad"}) or Gini's mean difference
 #'   (\code{robust = "gmd"}) might be more robust measures of dispersion. For
 #'   the latter option, \CRANpkg{sjstats} needs to be installed.
+#'   \code{robust = "2sd"} divides the centered variables by two standard
+#'   deviations, following a suggestion by \emph{Gelman (2008)}, so the
+#'   rescaled input is comparable to binary variables.
 #'
 #' @inheritParams to_factor
 #' @inheritParams rec
 #'
-#' @return A vector with standardized or centered variables. If \code{x} is a
-#'         data frame, only the transformed variables will be returned.
+#' @return If \code{x} is a vector, returns a vector with standardized or
+#'   centered variables. If \code{x} is a data frame, for \code{append = TRUE},
+#'   \code{x} including the transformed variables as new columns is returned;
+#'   if \code{append = FALSE}, only the transformed variables will be returned.
 #'
 #' @note \code{std()} and \code{center()} only return a vector, if \code{x} is
-#'         a vector. If \code{x} is a data frame and only one variable is specified
-#'         in the \code{...}-ellipses argument, both functions do return a
-#'         data frame (see 'Examples').
+#'   a vector. If \code{x} is a data frame and only one variable is specified
+#'   in the \code{...}-ellipses argument, both functions do return a
+#'   data frame (see 'Examples').
 #'
 #' @details \code{std()} and \code{center()} also work on grouped data frames
-#'          (see \code{\link[dplyr]{group_by}}). In this case, standardization
-#'          or centering is applied to the subsets of variables in \code{x}.
-#'          See 'Examples'.
+#'   (see \code{\link[dplyr]{group_by}}). In this case, standardization
+#'   or centering is applied to the subsets of variables in \code{x}.
+#'   See 'Examples'.
+#'   \cr \cr
+#'   For more complicated models with many predictors, Gelman and Hill (2007)
+#'   suggest leaving binary inputs as is and only standardize continuous predictors
+#'   by dividing by two standard deviations. This ensures a rough comparability
+#'   in the coefficients.
+#'
+#' @references
+#'   Gelman A (2008) Scaling regression inputs by dividing by two
+#'   standard deviations. \emph{Statistics in Medicine 27: 2865–2873.}
+#'   \url{http://www.stat.columbia.edu/~gelman/research/published/standardizing7.pdf}
+#'   \cr \cr
+#'   Gelman A, Hill J (2007) Data Analysis Using Regression and Multilevel/Hierarchical
+#'   Models. Cambdridge, Cambdrige University Press: 55-57
 #'
 #' @examples
 #' data(efc)
@@ -51,6 +73,7 @@
 #' # works also with grouped data frames
 #' mtcars %>% std(disp)
 #'
+#' # compare new column "disp_z" w/ output above
 #' mtcars %>%
 #'   group_by(cyl) %>%
 #'   std(disp)
@@ -61,11 +84,40 @@
 #' # don't standardize factors
 #' std(iris, include.fac = FALSE, append = FALSE)
 #'
+#' # standardize only variables with more than 10 unique values
+#' p <- function(x) dplyr::n_distinct(x) > 10
+#' std_if(efc, predicate = p, append = FALSE)
+#'
 #' @importFrom dplyr quos
 #' @export
-std <- function(x, ..., robust = c("sd", "gmd", "mad"), include.fac = FALSE, append = TRUE, suffix = "_z") {
+std <- function(x, ..., robust = c("sd", "2sd", "gmd", "mad"), include.fac = FALSE, append = TRUE, suffix = "_z") {
   # evaluate arguments, generate data
   .dat <- get_dot_data(x, dplyr::quos(...))
+
+  # match arguments
+  robust <- match.arg(robust)
+
+  std_and_center(x, .dat, include.fac, append, standardize = TRUE, robust = robust, suffix)
+}
+
+
+#' @importFrom dplyr select_if
+#' @rdname std
+#' @export
+std_if <- function(x, predicate, robust = c("sd", "2sd", "gmd", "mad"), include.fac = FALSE, append = TRUE, suffix = "_z") {
+
+  # select variables that match logical conditions
+  .dat <- dplyr::select_if(x, .predicate = predicate)
+
+  # if no variable matches the condition specified
+  # in predicate, return original data
+
+  if (sjmisc::is_empty(.dat)) {
+    if (append)
+      return(x)
+    else
+      return(.dat)
+  }
 
   # match arguments
   robust <- match.arg(robust)
@@ -79,6 +131,28 @@ std <- function(x, ..., robust = c("sd", "gmd", "mad"), include.fac = FALSE, app
 center <- function(x, ..., include.fac = FALSE, append = TRUE, suffix = "_c") {
   # evaluate arguments, generate data
   .dat <- get_dot_data(x, dplyr::quos(...))
+
+  std_and_center(x, .dat, include.fac, append, standardize = FALSE, robust = NULL, suffix)
+}
+
+
+#' @importFrom dplyr select_if
+#' @rdname std
+#' @export
+center_if <- function(x, predicate, include.fac = FALSE, append = TRUE, suffix = "_c") {
+
+  # select variables that match logical conditions
+  .dat <- dplyr::select_if(x, .predicate = predicate)
+
+  # if no variable matches the condition specified
+  # in predicate, return original data
+
+  if (sjmisc::is_empty(.dat)) {
+    if (append)
+      return(x)
+    else
+      return(.dat)
+  }
 
   std_and_center(x, .dat, include.fac, append, standardize = FALSE, robust = NULL, suffix)
 }
@@ -104,7 +178,7 @@ std_helper <- function(x, include.fac, standardize, robust) {
   # check whether factors should also be standardized
   if (is.factor(x)) {
     if (include.fac)
-      x <- to_value(x, keep.labels = FALSE)
+      x <- sjlabelled::as_numeric(x, keep.labels = FALSE)
     else
       return(x)
   }
@@ -129,6 +203,8 @@ std_helper <- function(x, include.fac, standardize, robust) {
       tmp <- tmp / stats::mad(tmp)
     else if (robust == "gmd" && requireNamespace("sjstats", quietly = TRUE))
       tmp <- tmp / sjstats::gmd(tmp)
+    else if (robust == "2sd")
+      tmp <- tmp / (2 * stats::sd(tmp))
     else
       tmp <- tmp / stats::sd(tmp)
   }