+++ NEWS.md
@@ -1,3 +1,34 @@
+# sjmisc 2.7.1
+
+## General
+
+* Fix typos and revise outdated paragraphs in vignettes.
+
+## New functions
+
+The recoding and transformation functions get scoped variants, allowing to select variables based on logical conditions described in a function:
+
+* `rec_if()` as scoped variant of `rec()`.
+* `dicho_if()` as scoped variant of `dicho()`.
+* `center_if()` as scoped variant of `center()`.
+* `std_if()` as scoped variant of `std()`.
+* `split_var_if()` as scoped variant of `split_var()`.
+* `group_var_if()` and `group_label_if()` as scoped variant of `group_var()` and `group_label()`.
+* `recode_to_if()` as scoped variant of `recode_to()`.
+* `set_na_if()` as scoped variant of `set_na()`.
+
+## Changes to functions
+
+* New function `remove_cols()` as alias for `remove_var()`.
+* `std()` gets a new robust-option, `robust = "2sd"`, which divides the centered variables by two standard deviations.
+* Slightly improved performance for `set_na()`.
+
+## Bug fixes
+
+* `frq()` now removes empty columns before computing frequencies, because applying `frq()` on empty vectors caused an error.
+* `empty_cols()` and `empty_rows()` (and hence, `remove_empty_cols()` and `remove_empty_rows()`) caused an error for data frames with only one column resp. row, or if `x` was a vector and no data frame.
+* `frq()` now removes missing values from input when weights are applied, to ensure that input and weights have same length.
+
 # sjmisc 2.7.0
 
 ## General